refactor parser from object, implement PandasPDB util methods
