+++ biopandas/mmcif/pandas_mmcif.py
@@ -1,4 +1,4 @@
-""" Class for working with MMCIF files."""
+"""Class for working with MMCIF files."""
 
 # BioPandas
 # Author: Arian Jamasb <arian@jamasb.io>, Sebastian Raschka <mail@sebastianraschka.com>
@@ -11,29 +11,11 @@
 #   See https://gitlab.com/pdbjapan/tools/cif-parsers/blob/master/LICENSE
 
 import gzip
-import os
-import re
-
-try:
-    import json
-except:
-    import simplejson as json
-
-try:
-    str.partition
-
-    def partitionString(string, sep):
-        return string.partition(sep)
-
-except:
-
-    def partitionString(string, sep):
-        tmp = string.split(sep)
-        return [tmp.pop(0), sep, sep.join(tmp)]
-
-
 import sys
-from typing import Dict, Optional
+from typing import Dict
+
+import numpy as np
+import pandas as pd
 
 try:
     from urllib.error import HTTPError, URLError
@@ -41,16 +23,24 @@ try:
 except ImportError:
     from urllib2 import urlopen, HTTPError, URLError  # Python 2.7 compatible
 
-import pandas as pd
+import warnings
+from distutils.version import LooseVersion
+
+from ..pdb.engines import amino3to1dict
+from .engines import mmcif_col_types
+from .mmcif_parser import __loadCIFData__
+
+pd_version = LooseVersion(pd.__version__)
 
 
 class PandasMMCIF:
-    def __init__(self):
+    def __init__(self, use_auth: bool = True):
         self._df = None
         self.mmcif_text = ""
         self.header = ""
         self.code = ""
         self.mmcif_path = ""
+        self.auth = use_auth
 
     @property
     def df(self):
@@ -80,11 +70,11 @@ class PandasMMCIF:
 
         """
         self.mmcif_path, self.pdb_text = self._read_mmcif(path=path)
-        self._df = self._construct_df(text=self.pdb_text.splitlines(True))
+        self._df = self._construct_df(text=self.pdb_text)
         # self.header, self.code = self._parse_header_code() #TODO: implement
         return self
 
-    def fetch_mmcif(self, pdb_code):
+    def fetch_mmcif(self, pdb_code: str):
         """Fetches mmcif file contents from the Protein Databank at rcsb.org.
 
         Parameters
@@ -98,7 +88,7 @@ class PandasMMCIF:
 
         """
         self.mmcif_path, self.mmcif_text = self._fetch_mmcif(pdb_code)
-        self._df = self._construct_df(pdb_lines=self.mmcif_text)
+        self._df = self._construct_df(text=self.mmcif_text)
         return self
 
     def _construct_df(self, text: str):
@@ -107,6 +97,7 @@ class PandasMMCIF:
         self.data = data
         df: Dict[str, pd.DataFrame] = {}
         full_df = pd.DataFrame.from_records(data["atom_site"])
+        full_df = full_df.astype(mmcif_col_types, errors="ignore")
         df["ATOM"] = pd.DataFrame(full_df[full_df.group_PDB == "ATOM"])
         df["HETATM"] = pd.DataFrame(full_df[full_df.group_PDB == "HETATM"])
         return df
@@ -152,423 +143,272 @@ class PandasMMCIF:
             )
         return path, txt
 
+    def get(self, s, df=None, invert=False, records=("ATOM", "HETATM")):
+        """Filter PDB DataFrames by properties
 
-class _loop:
-    def __init__(self, parserObj):
-        self.parserObj = parserObj
-        self.length = 0
-        self.refID = -1
+        Parameters
+        ----------
+        s : str  in {'main chain', 'hydrogen', 'c-alpha', 'heavy'}
+            String to specify which entries to return.
 
-        self.refList = []
+        df : pandas.DataFrame, default: None
+            Optional DataFrame to perform the filter operation on.
+            If df=None, filters on self.df['ATOM'].
 
-        self.namesDefined = False
+        invert : bool, default: True
+            Inverts the search query. For example if s='hydrogen' and
+            invert=True, all but hydrogen entries are returned.
 
-    def addName(self, name):
-        catName = type(name) == str and partitionString(name, ".") or ["", "", ""]
-        if catName[1]:
-            if not catName[0] in self.parserObj.currentTarget[-2]:
-                self.parserObj.currentTarget[-2][catName[0]] = {}
-            if not catName[2] in self.parserObj.currentTarget[-2][catName[0]]:
-                self.parserObj.currentTarget[-2][catName[0]][catName[2]] = []
-            self.refList.append(
-                self.parserObj.currentTarget[-2][catName[0]][catName[2]]
+        records : iterable, default: ('ATOM', 'HETATM')
+            Specify which record sections to consider. For example, to consider
+            both protein and ligand atoms, set `records=('ATOM', 'HETATM')`.
+            This setting is ignored if `df` is not set to None.
+            For downward compatibility, a string argument is still supported
+            but deprecated and will be removed in future versions.
+
+        Returns
+        --------
+        df : pandas.DataFrame
+            Returns a DataFrame view on the filtered entries.
+
+        """
+        if isinstance(records, str):
+            warnings.warn(
+                "Using a string as `records` argument is "
+                "deprecated and will not be supported in future"
+                " versions. Please use a tuple or"
+                " other iterable instead",
+                DeprecationWarning,
             )
+            records = (records,)
+
+        if not self._get_dict:
+            self._get_dict = self._init_get_dict()
+        if s not in self._get_dict.keys():
+            raise AttributeError("s must be in %s" % self._get_dict.keys())
+        if not df:
+            df = pd.concat(objs=[self.df[i] for i in records])
+        return self._get_dict[s](df, invert=invert)
+
+    @staticmethod
+    def _get_mainchain(df, invert: bool = False, atom_col: str = "auth_atom_id"):
+        """Return only main chain atom entries from a DataFrame"""
+        return (
+            df[
+                (df[atom_col] != "C")
+                & (df[atom_col] != "O")
+                & (df[atom_col] != "N")
+                & (df[atom_col] != "CA")
+            ]
+            if invert
+            else df[
+                (df[atom_col] == "C")
+                | (df[atom_col] == "O")
+                | (df[atom_col] == "N")
+                | (df[atom_col] == "CA")
+            ]
+        )
+
+    @staticmethod
+    def _get_hydrogen(df, invert):
+        """Return only hydrogen atom entries from a DataFrame"""
+        if invert:
+            return df[(df["type_symbol"] != "H")]
         else:
-            if not catName[0] in self.parserObj.currentTarget[-2]:
-                self.parserObj.currentTarget[-2][catName[0]] = []
-            self.refList.append(self.parserObj.currentTarget[-2][catName[0]])
-        self.length = len(self.refList)
+            return df[(df["type_symbol"] == "H")]
 
-    def pushValue(self, value):
-        if not self.namesDefined:
-            self.namesDefined = True
-        target = self.nextTarget()
-        if value == "stop_":
-            return self.stopPush()
-        target.append(value)
+    @staticmethod
+    def _get_heavy(df, invert):
+        """Return only heavy atom entries from a DataFrame"""
+        if invert:
+            return df[df["type_symbol"] == "H"]
+        else:
+            return df[df["type_symbol"] != "H"]
 
-    def nextTarget(self):
-        self.refID = (self.refID + 1) % self.length
-        return self.refList[self.refID]
+    @staticmethod
+    def _get_calpha(df, invert, atom_col: str = "auth_atom_id"):
+        """Return c-alpha atom entries from a DataFrame"""
+        return df[df[atom_col] != "CA"] if invert else df[df[atom_col] == "CA"]
 
-    def stopPush(self):
-        self.refID = -1
+    @staticmethod
+    def _get_carbon(df, invert):
+        """Return carbon atom entries from a DataFrame"""
+        if invert:
+            return df[df["type_symbol"] == "C"]
+        else:
+            return df[df["type_symbol"] != "C"]
 
+    def amino3to1(
+        self,
+        record: str = "ATOM",
+        residue_col: str = "auth_comp_id",
+        residue_number_col: str = "auth_seq_id",
+        chain_col: str = "auth_asym_id",
+        fillna: str = "?",
+    ):
+        """Creates 1-letter amino acid codes from DataFrame
 
-def specialSplit(content):
-    output = [["", False]]
-    quote = False
-    length = len(content)
-    log = ""
-    for c in range(length):
-        isWS = content[c] in [" ", "\t"]
-        if content[c] in ["'", '"'] and (
-            (
-                c == 0
-                or content[c - 1] == " "
-                or content[c - 1] == "\t"
-                or c == length - 1
-                or content[c + 1] == " "
-                or content[c + 1] == "\t"
+        Non-canonical amino-acids are converted as follows:
+        ASH (protonated ASP) => D
+        CYX (disulfide-bonded CYS) => C
+        GLH (protonated GLU) => E
+        HID/HIE/HIP (different protonation states of HIS) = H
+        HYP (hydroxyproline) => P
+        MSE (selenomethionine) => M
+
+        Parameters
+        ----------
+        record : str, default: 'ATOM'
+            Specfies the record DataFrame.
+        residue_col : str,  default: 'residue_name'
+            Column in `record` DataFrame to look for 3-letter amino acid
+            codes for the conversion.
+        fillna : str, default: '?'
+            Placeholder string to use for unknown amino acids.
+
+        Returns
+        ---------
+        pandas.DataFrame : Pandas DataFrame object consisting of two columns,
+            `'chain_id'` and `'residue_name'`, where the former contains
+            the chain ID of the amino acid and the latter
+            contains the 1-letter amino acid code, respectively.
+
+        """
+        tmp = self.df[record]
+        cmp = "placeholder"
+        indices = []
+
+        residue_number_insertion = (
+            tmp[residue_number_col].astype(str) + tmp["pdbx_PDB_ins_code"]
+        )
+
+        for num, ind in zip(residue_number_insertion, np.arange(tmp.shape[0])):
+            if num != cmp:
+                indices.append(ind)
+            cmp = num
+
+        transl = tmp.iloc[indices][residue_col].map(amino3to1dict).fillna(fillna)
+
+        return pd.concat((tmp.iloc[indices][chain_col], transl), axis=1)
+
+    @staticmethod
+    def rmsd(df1, df2, s=None, invert=False):
+        """Compute the Root Mean Square Deviation between molecules.
+
+        Parameters
+        ----------
+        df1 : pandas.DataFrame
+            DataFrame with HETATM, ATOM, and/or ANISOU entries.
+
+        df2 : pandas.DataFrame
+            Second DataFrame for RMSD computation against df1. Must have the
+            same number of entries as df1.
+
+        s : {'main chain', 'hydrogen', 'c-alpha', 'heavy', 'carbon'} or None,
+            default: None
+            String to specify which entries to consider. If None, considers
+            all atoms for comparison.
+
+        invert : bool, default: False
+            Inverts the string query if true. For example, the setting
+            `s='hydrogen', invert=True` computes the RMSD based on all
+            but hydrogen atoms.
+
+        Returns
+        ---------
+        rmsd : float
+            Root Mean Square Deviation between df1 and df2
+
+        """
+        if df1.shape[0] != df2.shape[0]:
+            raise AttributeError("DataFrames have unequal lengths")
+        get_dict = PandasMMCIF._init_get_dict()
+        if s:
+            if s not in get_dict.keys():
+                raise AttributeError("s must be in " "%s or None" % get_dict.keys())
+            df1 = get_dict[s](df1, invert=invert)
+            df2 = get_dict[s](df2, invert=invert)
+
+        total = (
+            (df1["Cartn_x"].values - df2["Cartn_x"].values) ** 2
+            + (df1["Cartn_y"].values - df2["Cartn_y"].values) ** 2
+            + (df1["Cartn_z"].values - df2["Cartn_z"].values) ** 2
+        )
+        return round((total.sum() / df1.shape[0]) ** 0.5, 4)
+
+    def distance(self, xyz=(0.00, 0.00, 0.00), records=("ATOM", "HETATM")):
+        """Computes Euclidean distance between atoms and a 3D point.
+
+        Parameters
+        ----------
+        xyz : tuple, default: (0.00, 0.00, 0.00)
+            X, Y, and Z coordinate of the reference center for the distance
+            computation.
+        records : iterable, default: ('ATOM', 'HETATM')
+            Specify which record sections to consider. For example, to consider
+            both protein and ligand atoms, set `records=('ATOM', 'HETATM')`.
+            This setting is ignored if `df` is not set to None.
+            For downward compatibility, a string argument is still supported
+            but deprecated and will be removed in future versions.
+
+        Returns
+        ---------
+        pandas.Series : Pandas Series object containing the Euclidean
+            distance between the atoms in the record section and `xyz`.
+
+        """
+
+        if isinstance(records, str):
+            warnings.warn(
+                "Using a string as `records` argument is "
+                "deprecated and will not be supported in future"
+                " versions. Please use a tuple or"
+                " other iterable instead",
+                DeprecationWarning,
             )
-        ):
-            quote = not quote
-        elif not quote and isWS and output[-1][0] != "":
-            output.append(["", False])
-        elif not quote and content[c] == "#":
-            break
-        elif not isWS or quote:
-            output[-1][0] += content[c]
-            output[-1][1] = quote
-    if output[-1][0] == "":
-        output.pop()
-    return output
+            records = (records,)
 
+        df = pd.concat(objs=[self.df[i] for i in records])
 
-class targetSetter:
-    def __init__(self, obj, key):
-        self.obj = obj
-        self.key = key
+        return np.sqrt(
+            np.sum(
+                df[["Cartn_x", "Cartn_y", "Cartn_z"]].subtract(xyz, axis=1) ** 2, axis=1
+            )
+        )
 
-    def setValue(self, value):
-        self.obj[self.key] = value
+    @staticmethod
+    def distance_df(df, xyz=(0.00, 0.00, 0.00)):
+        """Computes Euclidean distance between atoms and a 3D point.
 
+        Parameters
+        ----------
+        df : DataFrame
+            DataFrame containing entries in the `PandasPdb.df['ATOM']`
+            or `PandasPdb.df['HETATM']` format for the
+            the distance computation to the `xyz` reference coordinates.
+        xyz : tuple, default: (0.00, 0.00, 0.00)
+            X, Y, and Z coordinate of the reference center for the distance
+            computation.
 
-class CIFparser:
-    def __init__(self):
-        self.data = {}
-        self.currentTarget = None
-        self.loopPointer = None
+        Returns
+        ---------
+        pandas.Series : Pandas Series object containing the Euclidean
+            distance between the atoms in the record section and `xyz`.
 
-    def parseString(self, contents):
-        multi_line_mode = False
-        buffer = []
-        for line in contents.splitlines():
-            Z = line[:1]
-            line = line.strip()
-            if Z == ";":
-                if multi_line_mode:
-                    self.setDataValue("\n".join(buffer))
-                else:
-                    buffer = []
-                multi_line_mode = not multi_line_mode
-                line = line[1:].strip()
-            if multi_line_mode:
-                buffer.append(line)
-            else:
-                self.processContent(specialSplit(line))
+        """
+        return np.sqrt(
+            np.sum(
+                df[["Cartn_x", "Cartn_y", "Cartn_z"]].subtract(xyz, axis=1) ** 2, axis=1
+            )
+        )
 
-    def parse(self, fileobj):
-        multi_line_mode = False
-        buffer = []
-        for line in fileobj.readlines():
-            Z = line[:1]
-            line = line.strip()
-            if Z == ";":
-                if multi_line_mode:
-                    self.setDataValue("\n".join(buffer))
-                else:
-                    buffer = []
-                multi_line_mode = not multi_line_mode
-                line = line[1:].strip()
-            if multi_line_mode:
-                buffer.append(line)
-            else:
-                self.processContent(specialSplit(line))
-
-    def processContent(self, content):
-        for c, quoted in content:
-            if c == "global_" and not quoted:
-                self.loopPointer = None
-                self.selectGlobal()
-            elif c[:5] == "data_" and not quoted:
-                self.loopPointer = None
-                self.selectData(c)
-            elif c[:5] == "save_" and not quoted:
-                self.loopPointer = None
-                if c[5:]:
-                    self.selectFrame(c)
-                else:
-                    self.endFrame()
-            elif c == "loop_" and not quoted:
-                self.loopPointer = _loop(self)
-            elif c[:1] == "_" and not quoted:
-                self.setDataName(c[1:])
-            else:
-                self.setDataValue(c)
-
-    def setDataName(self, name):
-        if self.loopPointer != None:
-            if self.loopPointer.namesDefined:
-                self.loopPointer = None
-            else:
-                return self.loopPointer.addName(name)
-        name = partitionString(name, ".")
-        self.currentTarget.pop()
-        if name[1]:
-            if not name[0] in self.currentTarget[-1]:
-                self.currentTarget[-1][name[0]] = {}
-            self.currentTarget[-1][name[0]][name[2]] = ""
-            self.currentTarget = self.currentTarget + [
-                targetSetter(self.currentTarget[-1][name[0]], name[2])
-            ]
-        else:
-            self.currentTarget[-1][name[0]] = ""
-            self.currentTarget = self.currentTarget + [
-                targetSetter(self.currentTarget[-1], name[0])
-            ]
-
-    def setDataValue(self, value):
-        if self.loopPointer is None:
-            self.currentTarget[-1].setValue([value])
-
-        else:
-            self.loopPointer.pushValue(value)
-
-    def selectGlobal(self):
-        self.currentTarget = [self.data, self.data, None]
-
-    def selectData(self, name):
-        if name not in self.data:
-            self.data[name] = {}
-        self.currentTarget = [self.data, self.data[name], None]
-
-    def selectFrame(self, name=""):
-        if name not in self.currentTarget[1]:
-            self.currentTarget[1][name] = {}
-        self.currentTarget = self.currentTarget[:2] + [
-            self.currentTarget[1][name],
-            None,
-        ]
-
-    def endData(self):
-        self.currentTarget = self.currentTarget[:2]
-
-    def endFrame(self):
-        self.currentTarget = self.currentTarget[:3]
-
-
-####################################################################################################################################################
-
-
-class __CIFfloat__(float):
-    def __repr__(self):
-        return "%.15g" % self
-
-
-class __CIFint__(int):
-    def __repr__(self):
-        return str(self)
-
-
-def __CIFfloatRange__(inp):
-    try:
-        pos = inp.index("-", 1)
-        return (__CIFfloat__(inp[:pos]), __CIFfloat__(inp[pos + 1 :]))
-    except:
-        return (__CIFfloat__(inp),)
-
-
-def __CIFintRange__(inp):
-    try:
-        pos = inp.index("-", 1)
-        return (__CIFint__(inp[:pos]), __CIFint__(inp[pos + 1 :]))
-    except:
-        return (__CIFint__(inp),)
-
-
-def __loadCIFdic__(dicFile, force=False):
-    jsfDic = f"{dicFile[:-4]}.json"
-    jsf = f"{dicFile[:-4]}_summary.json"
-    dic = {}
-    try:
-        if force:
-            throw
-        dic = json.loads(open(jsf).read())
-    except:
-        parser = CIFparser()
-        parser.parse(open(dicFile))
-        json.dump(parser.data, open(jsfDic, "w"))
-        for k, v in parser.data["data_mmcif_pdbx.dic"].items():
-            if type(v) != dict or "item_type" not in v:
-                continue
-            name = partitionString(k[6:], ".")
-            if name[0] not in dic:
-                dic[name[0]] = {}
-            dic[name[0]][name[2]] = v["item_type"]["code"][0].strip()
-        json.dump(dic, open(jsf, "w"))
-
-    typing = {}
-    for k, v in dic.items():
-        for k2, v2 in v.items():
-            if v2 == "float":
-                if k not in typing:
-                    typing[k] = {}
-                typing[k][k2] = __CIFfloat__
-            elif v2 == "float-range":
-                if k not in typing:
-                    typing[k] = {}
-                typing[k][k2] = __CIFfloatRange__
-            elif v2 == "int":
-                if k not in typing:
-                    typing[k] = {}
-                typing[k][k2] = __CIFint__
-            elif v2 == "int-range":
-                if k not in typing:
-                    typing[k] = {}
-                typing[k][k2] = __CIFintRange__
-    return typing
-
-
-def __dumpCIF__(jso):
-    return __dumpPart__(jso)
-
-
-__cifStrCheck__ = re.compile(r"[\\s\(\)]")
-__cifStrNLCheck__ = re.compile(r"[\n]")
-
-
-def __dumpStr__(inp):
-    if inp == None:
-        return "?"
-    else:
-        if type(inp) != str and type(inp) != unicode:
-            return str(inp)
-        if re.search(__cifStrNLCheck__, inp) != None:
-            return "\n;%s\n;" % inp
-        if re.search(__cifStrCheck__, inp) != None:
-            return "'%s'" % inp
-        else:
-            return inp
-
-
-def __padString__(inp, flength):
-    return inp + (" " * (flength - len(inp)))
-
-
-def __dumpCat__(k, v):
-    output = "#\n"
-    noi = len(v[v.keys()[0]])
-    if noi == 1:
-        pad = 0
-        for k2 in v.keys():
-            if len(k2) > pad:
-                pad = len(k2)
-        pad += 3
-        for k2 in v.keys():
-            output += "_%s.%s%s\n" % (k, __padString__(k2, pad), __dumpStr__(v[k2][0]))
-    else:
-        output += "loop_\n"
-        pad = []
-        for k2 in v.keys():
-            output += "_%s.%s\n" % (k, k2)
-            pad.append(0)
-        tmp1 = []
-        for i in range(noi):
-            tmp2 = []
-            tmp1.append(tmp2)
-            for k2 in v.keys():
-                tmp2.append(__dumpStr__(v[k2][i]))
-
-        for j in range(len(tmp1[0])):
-            pad = 0
-            for i in range(len(tmp1)):
-                if tmp1[i][j][:2] != "\n;" and len(tmp1[i][j]) > pad:
-                    pad = len(tmp1[i][j])
-            pad += 1
-            for i in range(len(tmp1)):
-                if tmp1[i][0][:2] != "\n;":
-                    tmp1[i][j] = __padString__(tmp1[i][j], pad)
-
-        for i in range(noi):
-            output += "".join(tmp1[i]) + "\n"
-
-    return output.strip() + "\n"
-
-
-def __dumpPart__(jso):
-    inner = True
-    output = ""
-    for k, v in jso.items():
-        if isinstance(v, dict):
-            if k[:5] != "data_" and k[:5] != "save_" and k[:7] != "global_":
-                output += __dumpCat__(k, v)
-            else:
-                output += k + "\n"
-                output += __dumpPart__(v)
-                inner = False
-    if inner:
-        return output + "#\n"
-    else:
-        return output
-
-
-def __loadCIFData__(data, doClean=True, doType=True):
-    parser = CIFparser()
-    if type(data) == str:
-        parser.parseString(data)
-    else:
-        parser.parse(data)  # fileobj
-
-    if not doClean:
-        return parser.data
-
-    for k, v in parser.data.items():
-        for k2, v2 in v.items():
-            for k3, v3 in v2.items():
-                for i in range(len(v3)):
-                    v2[k3][i] = v3[i] not in ["?", "."] and v3[i] or None
-
-    if not doType or not __mmcifTyping__:
-        return parser.data
-
-    for struct, data in parser.data.items():
-        for k, v in __mmcifTyping__.items():
-            if k not in data:
-                continue
-            for k2, v2 in v.items():
-                if k2 in data[k]:
-                    for r in range(len(data[k][k2])):
-                        try:
-                            data[k][k2][r] = v2(data[k][k2][r])
-                        except:
-                            pass
-
-    return parser.data
-
-
-def __loadCIF__(cifFile, doClean=True, doType=True):
-    parser = CIFparser()
-    if cifFile[-3:].lower() == ".gz":
-        parser.parse(gzip.open(cifFile))
-    else:
-        parser.parse(open(cifFile))
-
-    if not doClean:
-        return parser.data
-
-    for k, v in parser.data.items():
-        for k2, v2 in v.items():
-            for k3, v3 in v2.items():
-                for i in range(len(v3)):
-                    v2[k3][i] = v3[i] not in ["?", "."] and v3[i] or None
-
-    if not doType or not __mmcifTyping__:
-        return parser.data
-
-    for struct, data in parser.data.items():
-        for k, v in __mmcifTyping__.items():
-            if k not in data:
-                continue
-            for k2, v2 in v.items():
-                if k2 in data[k]:
-                    for r in range(len(data[k][k2])):
-                        try:
-                            data[k][k2][r] = v2(data[k][k2][r])
-                        except:
-                            pass
-
-    return parser.data
-
-
-__mmcifTyping__ = None
+    @staticmethod
+    def _init_get_dict():
+        """Initialize dictionary for filter operations."""
+        return {
+            "main chain": PandasMMCIF._get_mainchain,
+            "hydrogen": PandasMMCIF._get_hydrogen,
+            "c-alpha": PandasMMCIF._get_calpha,
+            "carbon": PandasMMCIF._get_carbon,
+            "heavy": PandasMMCIF._get_heavy,
+        }
